Write_In_Air-master/Write_In_Air-master/requirements.txt
absl-py==0.9.0
astor==0.8.1
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-models==0.0.7
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
murmurhash==1.0.2
networkx==2.4
numpy==1.18.4
opencv-python==3.4.5.20
pathlib==1.0.1
Pillow==7.1.2
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
protobuf==3.12.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
six==1.15.0
spacy==2.2.4
srsly==1.0.2
tensorboard==1.14.0
tensorflow==1.15.2
tensorflow-estimator==1.14.0
termcolor==1.1.0
thinc==7.4.0
tqdm==4.46.0
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
